Fix bug where nodes would block indefinitely during download
In testing on a large cluster, it was noticed that nodes would
occasionally become non-responsive after a deployment. The cause was
found to be that the file share task would block attempting to send a
packet to the downloader for a file if a file part was recieved after
all parts were downloaded for that file, but before the download task
was concluded.